Apply IAM patches also on startup of STS, prevent patching multiple times
